+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/actions/FlexibleSearchOpenSettingsAction.kt
@@ -18,17 +18,17 @@
 
 package com.intellij.idea.plugin.hybris.flexibleSearch.actions
 
+import com.intellij.icons.AllIcons
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
-import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.flexibleSearch.settings.FlexibleSearchSettingsConfigurableProvider
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.options.ShowSettingsUtil
 
-class OpenFlexibleSearchSettingsAction : AnAction(
+class FlexibleSearchOpenSettingsAction : AnAction(
     message("hybris.fxs.actions.open_settings"),
     message("hybris.fxs.actions.open_settings.description"),
-    HybrisIcons.FLEXIBLE_SEARCH
+    AllIcons.General.Settings
 ) {
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return